final version

+++ src/components/jobDisplay/jobDetail.js
@@ -1,7 +1,5 @@
-import React, { Component, useEffect, useState } from "react";
+import React, { useEffect, useState } from "react";
 import './jobDetail.css';
-import jsonData from './testdata.json';
-import { Link } from "react-router-dom";
 import axios from "axios";
 
 export default function JobDetail() {
@@ -17,7 +15,8 @@ export default function JobDetail() {
         location: '',
         description: '',
         email: '',
-        website: ''
+        website: '',
+        birthday: ''
     });
 
     const [jobData, setJob] = useState({
@@ -36,8 +35,6 @@ export default function JobDetail() {
             .then(resUser => {
                 if (resUser.data !== '') {
                     setIsLog('is-log');
-                    // console.log(resUser.data)
-                    const name = resUser.data;
                     setLog(resUser.data);
                     axios.get('/api/jobs/findId/' + id)
                         .then(response => {
@@ -50,7 +47,8 @@ export default function JobDetail() {
                                 location: response.data[0].location,
                                 description: response.data[0].description,
                                 email: response.data[0].email,
-                                website: response.data[0].website
+                                website: response.data[0].website,
+                                birthday: response.data[0].birthday
                             })
                             if (resUser.data === response.data[0].user) {
                                 setHost('is-host');
@@ -75,8 +73,6 @@ export default function JobDetail() {
         return errorUser;
     }
 
-
-
     function getJobDetail() {
         var url = '/api/jobs/findId/' + id;
         axios.get(url)
@@ -91,19 +87,17 @@ export default function JobDetail() {
             });
 
         getLoggedUser().catch(errorUser => console.error(errorUser));
-        console.log(userLog);
-
-
     }
 
 
     useEffect(getJobDetail, []);
 
     const handleFavorite = () => {
-        console.log(jobIfo);
+        if (isLogged === false) {
+            window.location.href = '/sign-in';
+        }
 
         if (isFav === true) {
-            console.log(isFav)
             var urlUser = '/api/users/unfavorite';
             var urlFav = '/api/favorites/delete/' + jobData.datas._id + '/' + userLog;
             axios.delete(urlFav)
@@ -135,39 +129,40 @@ export default function JobDetail() {
 
     const handleDelete = () => {
         var urlFav = '/api/favorites/deleteJob/' + jobIfo.jobid;
-        console.log(urlFav)
         var urlJob = '/api/jobs/delete/' + id;
         console.log(urlJob)
-        axios.delete(urlFav)
-        .then(response => {
-            // window.location.href = '/';
-            console.log(response)
-        })
-        .catch(error => console.error(error));
+        if (isFav === true) {
+            axios.delete(urlFav)
+                .then(response => {
+                    console.log(response)
+                })
+                .catch(error => {
+                    console.error(error)
+                    alert("Delete fail");
+                });
+        }
+
         axios.delete(urlJob)
             .then(response => {
-                // window.location.href = '/';
                 console.log(response)
-            })
-            .catch(error => console.error(error));
-        
+                window.location.href = '/';
 
+            })
+            .catch(error => {
+                console.error(error)
+                alert("Delete fail");
+            });
     }
 
     const handleEdit = () => {
-        window.location.href = '/editjob/' + id+'/'+isFav;
+        window.location.href = '/editjob/' + id + '/' + isFav;
     }
-
-
 
     return <div>
         <div className="outer">
             <div className="inner">
-                {/* <form> */}
                 <div className="mb-3 row">
-                    {/* <label for="staticTitle" className="col-sm-12 col-form-label">Job Title:  {jobData.datas.job}</label> */}
                     <label for="staticTitle" className="col-sm-12 col-form-label">Job Title:  {jobData.datas.job}</label>
-
                 </div>
                 <div className="mb-3 row">
                     <label for="staticCompany" className="col-sm-12 col-form-label">Company:  {jobData.datas.company}</label>
@@ -175,9 +170,6 @@ export default function JobDetail() {
                 <div className="mb-3 row">
                     <label for="staticLocation" className="col-sm-12 col-form-label">Location:  {jobData.datas.location}</label>
                 </div>
-                {/* <div className="mb-3 row">
-                            <label for="staticDate" className="col-sm-12 col-form-label">Post Date:  {jsonData.date}</label>
-                        </div> */}
                 <div className="mb-3 row">
                     <label for="staticDescription" className="col-sm-12 col-form-label">Description:  {jobData.datas.description}</label>
                 </div>
@@ -185,7 +177,10 @@ export default function JobDetail() {
                     <label for="staticEmail" className="col-sm-12 col-form-label">Contact Email:  {jobData.datas.email}</label>
                 </div>
                 <div className="mb-3 row">
-                    <label for="staticWeb" className="col-sm-12 col-form-label">Webpage:  {jobData.datas.website}</label>
+                    <label for="staticWeb" className="col-sm-12 col-form-label">Website:  {jobData.datas.website}</label>
+                </div>
+                <div className="mb-3 row">
+                    <label for="staticWeb" className="col-sm-12 col-form-label">Post Date:  {jobData.datas.birthday}</label>
                 </div>
                 <div className="mb-3">
                     <div className="col-sm-offset-2 col-sm-2">
@@ -198,9 +193,7 @@ export default function JobDetail() {
                         <button type="button" className="btn btn-info btn-lg" id={isHost} onClick={handleDelete}>Delete</button>
                     </div>
                 </div>
-                {/* </form> */}
             </div>
         </div >
     </div>
-
 }
